sales.java, yourproduct.java, product.java, buy.java, marketplaceservice.java, sales.jsp, viewkatalog.jsp, purchase.jspyourproduct.jsp

+++ MarketplaceService/src/java/MarketplaceService/product.java
@@ -13,10 +13,10 @@ public class product {
     private String usernama;
     private String imgpath;
     private String namabarang;
-    private double hargabarang;
+    private long hargabarang;
     private String deskripsi;
-    private int jumlahlike;
-    private int jumlahbeli;
+    private long jumlahlike;
+    private long jumlahbeli;
 
     public void setUsernama(String usernama) {
         this.usernama = usernama;
@@ -30,7 +30,7 @@ public class product {
         this.namabarang = namabarang;
     }
 
-    public void setHargabarang(double hargabarang) {
+    public void setHargabarang(long hargabarang) {
         this.hargabarang = hargabarang;
     }
 
@@ -38,11 +38,11 @@ public class product {
         this.deskripsi = deskripsi;
     }
 
-    public void setJumlahlike(int jumlahlike) {
+    public void setJumlahlike(long jumlahlike) {
         this.jumlahlike = jumlahlike;
     }
 
-    public void setJumlahbeli(int jumlahbeli) {
+    public void setJumlahbeli(long jumlahbeli) {
         this.jumlahbeli = jumlahbeli;
     }
 
@@ -58,7 +58,7 @@ public class product {
         return namabarang;
     }
 
-    public double getHargabarang() {
+    public long getHargabarang() {
         return hargabarang;
     }
 
@@ -66,17 +66,17 @@ public class product {
         return deskripsi;
     }
 
-    public int getJumlahlike() {
+    public long getJumlahlike() {
         return jumlahlike;
     }
 
-    public int getJumlahbeli() {
+    public long getJumlahbeli() {
         return jumlahbeli;
     }
 
     public product(String usernama, String imgpath, String namabarang,
-            double hargabarang, String deskripsi, int jumlahlike,
-            int jumlahbeli) {
+            long hargabarang, String deskripsi, long jumlahlike,
+            long jumlahbeli) {
         this.usernama = usernama;
         this.imgpath = imgpath;
         this.namabarang = namabarang;
